+++ LogVerdict.psm1
@@ -35,6 +35,7 @@ Export-ModuleMember -Function @(
     'Test-LogVerdictDatabase',
     'Export-LogVerdictReport',
     'Export-LogVerdictStandard',
+    'Watch-LogVerdict',
     'Show-LogVerdictReport',
     'Show-LogVerdictGui'
 )